Animations, floor, optimization and comments..
ExitLevel.cs:
"juntarTareas" function was deleted and replaced by openDoor, which does the same, but optimized.

controles.cs:
Comments where added, the functions are now explained. The update was polished, now it has only the functions. Now doorCheck is the one who opens the door through "openDoor".

Win Screen got an animation and a script, where the animation activates.